Try to reduce memory used during compilation
