+++ src/hictk/balance/common.hpp
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: MIT
 
+#pragma once
+
 #include <fmt/compile.h>
 #include <fmt/format.h>
 #include <spdlog/spdlog.h>
@@ -37,7 +39,7 @@
 namespace hictk::tools {
 
 template <typename BalanceConfig>
-static void write_weights_hic(
+inline void write_weights_hic(
     hic::internal::HiCFileWriter& hfw, const BalanceConfig& c,
     const phmap::flat_hash_map<std::uint32_t, balancing::Weights>& weights, bool force_overwrite) {
   for (const auto& [resolution, weights_] : weights) {
@@ -47,7 +49,7 @@ static void write_weights_hic(
 }
 
 template <typename BalanceConfig>
-static void write_weights_cooler(std::string_view uri, const BalanceConfig& c,
+inline void write_weights_cooler(std::string_view uri, const BalanceConfig& c,
                                  const balancing::Weights& weights,
                                  const std::vector<double>& variance,
                                  const std::vector<double>& scale) {
@@ -111,13 +113,13 @@ static void write_weights_cooler(std::string_view uri, const BalanceConfig& c,
 }  // NOLINT(clang-analyzer-cplusplus.NewDeleteLeaks)
 
 template <typename BalanceConfig>
-static void write_weights_cooler(std::string_view uri, const BalanceConfig& c,
+inline void write_weights_cooler(std::string_view uri, const BalanceConfig& c,
                                  const balancing::Weights& weights) {
   return write_weights_cooler(uri, c, weights, {-1}, {-1});
 }
 
 template <typename Balancer, typename BalanceConfig>
-static auto init_params(const BalanceConfig& c, const std::filesystem::path& tmpfile) ->
+inline auto init_params(const BalanceConfig& c, const std::filesystem::path& tmpfile) ->
     typename Balancer::Params {
   if constexpr (std::is_same_v<Balancer, balancing::ICE>) {
     return {c.tolerance, c.max_iters, c.masked_diags, c.min_nnz, c.min_count,
@@ -132,7 +134,7 @@ static auto init_params(const BalanceConfig& c, const std::filesystem::path& tmp
 }
 
 template <typename Balancer, typename BalanceConfig>
-static int balance_cooler(cooler::File& f, const BalanceConfig& c,
+inline int balance_cooler(cooler::File& f, const BalanceConfig& c,
                           const std::filesystem::path& tmp_dir) {
   if (!c.force && !c.stdout_ && f.has_normalization(c.name)) {
     throw std::runtime_error(
@@ -177,7 +179,7 @@ static int balance_cooler(cooler::File& f, const BalanceConfig& c,
 }
 
 template <typename Balancer, typename BalanceConfig>
-static int balance_hic(const BalanceConfig& c, const std::filesystem::path& tmp_dir) {
+inline int balance_hic(const BalanceConfig& c, const std::filesystem::path& tmp_dir) {
   const auto resolutions = hic::utils::list_resolutions(c.path_to_input);
   for (const auto& res : resolutions) {
     const hic::File f(c.path_to_input.string(), res);
@@ -226,7 +228,7 @@ static int balance_hic(const BalanceConfig& c, const std::filesystem::path& tmp_
 }
 
 template <typename Balancer, typename BalanceConfig>
-static int balance_multires_cooler(const BalanceConfig& c, const std::filesystem::path& tmp_dir) {
+inline int balance_multires_cooler(const BalanceConfig& c, const std::filesystem::path& tmp_dir) {
   const auto resolutions = cooler::utils::list_resolutions(c.path_to_input.string());
 
   for (const auto& res : resolutions) {
@@ -235,50 +237,6 @@ static int balance_multires_cooler(const BalanceConfig& c, const std::filesystem
     balance_cooler<Balancer>(clr, c, tmp_dir);
   }
   return 0;
-}
-
-int balance_subcmd(const BalanceICEConfig& c) {
-  SPDLOG_INFO(FMT_STRING("balancing using ICE ({})"), c.name);
-  const auto tmp_dir =
-      !c.in_memory ? std::make_unique<const internal::TmpDir>(c.tmp_dir, true) : nullptr;
-  const std::filesystem::path& tmp_dir_path = tmp_dir ? (*tmp_dir)() : "";
-
-  if (hic::utils::is_hic_file(c.path_to_input.string())) {
-    return balance_hic<balancing::ICE>(c, tmp_dir_path);
-  }
-  if (cooler::utils::is_multires_file(c.path_to_input.string())) {
-    return balance_multires_cooler<balancing::ICE>(c, tmp_dir_path);
-  }
-  auto clr = cooler::File(c.path_to_input.string());
-  return balance_cooler<balancing::ICE>(clr, c, tmp_dir_path);
-}
-
-int balance_subcmd(const BalanceSCALEConfig& c) {
-  SPDLOG_INFO(FMT_STRING("balancing using SCALE ({})"), c.name);
-  const auto tmp_dir =
-      !c.in_memory ? std::make_unique<const internal::TmpDir>(c.tmp_dir, true) : nullptr;
-  const std::filesystem::path& tmp_dir_path = tmp_dir ? (*tmp_dir)() : "";
-
-  if (hic::utils::is_hic_file(c.path_to_input.string())) {
-    return balance_hic<balancing::SCALE>(c, tmp_dir_path);
-  }
-  if (cooler::utils::is_multires_file(c.path_to_input.string())) {
-    return balance_multires_cooler<balancing::SCALE>(c, tmp_dir_path);
-  }
-  auto clr = cooler::File(c.path_to_input.string());
-  return balance_cooler<balancing::SCALE>(clr, c, tmp_dir_path);
-}
-
-int balance_subcmd(const BalanceVCConfig& c) {
-  SPDLOG_INFO(FMT_STRING("balancing using VC ({})"), c.name);
-  if (hic::utils::is_hic_file(c.path_to_input.string())) {
-    return balance_hic<balancing::VC>(c, "");
-  }
-  if (cooler::utils::is_multires_file(c.path_to_input.string())) {
-    return balance_multires_cooler<balancing::VC>(c, "");
-  }
-  auto clr = cooler::File(c.path_to_input.string());
-  return balance_cooler<balancing::VC>(clr, c, "");
 }
 
 }  // namespace hictk::tools
