+++ src/hictk/merge/merge.cpp
@@ -2,40 +2,18 @@
 //
 // SPDX-License-Identifier: MIT
 
+#include "./merge.hpp"
+
 #include <fmt/format.h>
+#include <fmt/std.h>
 #include <spdlog/spdlog.h>
 
-#include <cassert>
 #include <chrono>
-#include <cstdint>
 #include <filesystem>
 
-#include "hictk/file.hpp"
 #include "hictk/tools/config.hpp"
 
 namespace hictk::tools {
-
-static void merge_to_cool(const MergeConfig& c) {
-  SPDLOG_INFO(FMT_STRING("begin merging {} files into one .{} file..."), c.input_files.size(),
-              c.output_format);
-  if (c.count_type == "int") {
-    utils::merge_to_cool<std::int32_t>(c.input_files.begin(), c.input_files.end(),
-                                       c.output_file.string(), c.resolution, c.force, c.chunk_size,
-                                       10'000'000, c.compression_lvl);
-    return;
-  }
-  assert(c.count_type == "float");
-  utils::merge_to_cool<double>(c.input_files.begin(), c.input_files.end(), c.output_file.string(),
-                               c.resolution, c.force, c.chunk_size, 10'000'000, c.compression_lvl);
-}
-
-static void merge_to_hic(const MergeConfig& c) {
-  SPDLOG_INFO(FMT_STRING("begin merging {} files into one .{} file..."), c.input_files.size(),
-              c.output_format);
-  utils::merge_to_hic(c.input_files.begin(), c.input_files.end(), c.output_file.string(),
-                      c.resolution, c.tmp_dir, c.force, c.chunk_size, c.threads, c.compression_lvl,
-                      c.skip_all_vs_all_matrix);
-}
 
 int merge_subcmd(const MergeConfig& c) {
   const auto t0 = std::chrono::system_clock::now();
