+++ src/hictk/merge/merge.hpp
@@ -1,0 +1,15 @@
+// Copyright (C) 2023 Roberto Rossini <roberros@uio.no>
+//
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+#include "hictk/tools/config.hpp"
+
+namespace hictk::tools {
+
+void merge_to_cool(const MergeConfig& c);
+
+void merge_to_hic(const MergeConfig& c);
+
+}  // namespace hictk::tools
