+++ src/hictk/metadata/mcool.cpp
@@ -2,24 +2,21 @@
 //
 // SPDX-License-Identifier: MIT
 
-#pragma once
-
 #include <fmt/format.h>
 
 #include <filesystem>
 #include <string>
-#include <toml++/toml.hpp>
 #include <utility>
 #include <vector>
 
-#include "./common.hpp"
-#include "./cool.hpp"
+#include "./metadata.hpp"
 #include "hictk/bin_table.hpp"
 #include "hictk/cooler/multires_cooler.hpp"
+#include "hictk/tools/toml.hpp"
 
 namespace hictk::tools {
 
-[[nodiscard]] inline toml::table normalize_attribute_map(const cooler::MultiResAttributes& map,
+[[nodiscard]] static toml::table normalize_attribute_map(const cooler::MultiResAttributes& map,
                                                          const std::string& uri) {
   toml::table attributes;
 
@@ -35,9 +32,8 @@ namespace hictk::tools {
   return attributes;
 }
 
-[[nodiscard]] inline int print_mcool_metadata(const std::filesystem::path& p,
-                                              MetadataOutputFormat format, bool include_file_path,
-                                              bool recursive) {
+int print_mcool_metadata(const std::filesystem::path& p, MetadataOutputFormat format,
+                         bool include_file_path, bool recursive) {
   const cooler::MultiResFile mclr(p);
   auto attributes = normalize_attribute_map(mclr.attributes(), include_file_path ? p.string() : "");
   std::vector<std::pair<std::string, toml::table>> nested_attributes{};
