+++ src/hictk/metadata/scool.cpp
@@ -2,20 +2,18 @@
 //
 // SPDX-License-Identifier: MIT
 
-#pragma once
-
+#include <cassert>
 #include <filesystem>
 #include <string>
-#include <toml++/toml.hpp>
 
-#include "./common.hpp"
-#include "./cool.hpp"
+#include "./metadata.hpp"
 #include "hictk/bin_table.hpp"
 #include "hictk/cooler/singlecell_cooler.hpp"
+#include "hictk/tools/toml.hpp"
 
 namespace hictk::tools {
 
-[[nodiscard]] inline toml::table normalize_attribute_map(const cooler::SingleCellAttributes& map,
+[[nodiscard]] static toml::table normalize_attribute_map(const cooler::SingleCellAttributes& map,
                                                          const std::string& uri) {
   toml::table attributes;
 
@@ -49,9 +47,8 @@ namespace hictk::tools {
   return attributes;
 }
 
-[[nodiscard]] inline int print_scool_metadata(const std::filesystem::path& p,
-                                              MetadataOutputFormat format, bool include_file_path,
-                                              bool recursive) {
+int print_scool_metadata(const std::filesystem::path& p, MetadataOutputFormat format,
+                         bool include_file_path, bool recursive) {
   const cooler::SingleCellFile sclr(p);
   auto attributes = normalize_attribute_map(sclr.attributes(), include_file_path ? p.string() : "");
   std::vector<std::pair<std::string, toml::table>> nested_attributes{};
