+++ test/units/transformers/CMakeLists.txt
@@ -10,7 +10,15 @@ include(Catch)
 
 add_executable(hictk_transformers_tests)
 
-target_sources(hictk_transformers_tests PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/transformers_test.cpp")
+target_sources(
+  hictk_transformers_tests
+  PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/transformers_coarsen_pixels_test.cpp"
+          "${CMAKE_CURRENT_SOURCE_DIR}/transformers_join_genomic_coords_test.cpp"
+          "${CMAKE_CURRENT_SOURCE_DIR}/transformers_pixel_merger_test.cpp"
+          "${CMAKE_CURRENT_SOURCE_DIR}/transformers_stats_test.cpp"
+          "${CMAKE_CURRENT_SOURCE_DIR}/transformers_to_dataframe_test.cpp"
+          "${CMAKE_CURRENT_SOURCE_DIR}/transformers_to_dense_matrix_test.cpp"
+          "${CMAKE_CURRENT_SOURCE_DIR}/transformers_to_sparse_matrix_test.cpp")
 
 target_link_libraries(
   hictk_transformers_tests
